Scope the Red Alert enums
The named enums in ra/ are enum class with using enum; the unnamed enums
and the named constant sets are constexpr constants, and tables indexed
by an enum are base::EnumArray sized by magic_enum::enum_count. The
scoping found five defects, each fixed and commented at the site:
event.cc stored a target's RTTI kind instead of the target, teamtype.cc
indexed FormationName with the union sibling Data.Quarry, overlay.cc
reset zones with the MZONE_ enumerators instead of the kZoneFlag mask,
cell.cc summed BQuantity past its end, and house.cc indexed
HouseTriggers by heap ID rather than house type.

+++ src/ra/aadata.cc
@@ -69,7 +69,6 @@
 #include "ra/externs.h"
 #include "ra/heap.h"
 #include "ra/house.h"
-#include "ra/jshell.h"
 #include "ra/mapedit.h"
 #include "ra/object.h"
 #include "ra/type.h"
@@ -609,5 +608,5 @@ void AircraftTypeClass::Dimensions(int& width, int& height) const {
  * HISTORY: * 07/09/1996 JLB : Created. *
  *=============================================================================================*/
 AircraftTypeClass& AircraftTypeClass::As_Reference(AircraftType aircraft) {
-  return *AircraftTypes.Ptr(aircraft);
+  return *AircraftTypes.Ptr(static_cast<int>(aircraft));
 }
