+++ src/ra/abstract_type.cc
@@ -23,7 +23,7 @@ int AbstractTypeClass::Full_Name() const {
   // object type and ID. A negative return signals the caller to look up the
   // string in NameOverride rather than the normal text table.
   for (base::ssize index = 0; index < std::ssize(NameOverride); index++) {
-    if (NameIDOverride[index] == ((RTTI + 1) * 100) + ID) {
+    if (NameIDOverride[index] == ((static_cast<int>(RTTI) + 1) * 100) + ID) {
       return static_cast<int>(-(index + 1));
     }
   }
