+++ src/ra/coord.cc
@@ -91,7 +91,8 @@ COORDINATE Coord_Move(const COORDINATE start, const DirType facing,
                       const uint16_t distance) {
   auto x = static_cast<int16_t>(Coord_X(start));
   auto y = static_cast<int16_t>(Coord_Y(start));
-  base::MovePoint(x, y, facing, static_cast<int16_t>(distance));
+  base::MovePoint(x, y, static_cast<uint8_t>(facing),
+                  static_cast<int16_t>(distance));
   return XY_Coord(static_cast<LEPTON>(x), static_cast<LEPTON>(y));
 }
